optimizations

+++ src/math.rs
@@ -1,8 +1,7 @@
 pub fn image_compare(target: Vec<u8>, image: Vec<u8>) -> f64{
-    let mut mse: f64 = image.iter().zip(target).map(|(a, b)| ((*a as f64 - b as f64).powf(2.0)) as f64).sum();
+    let mut mse: f64 = image.iter().zip(target).map(|(a, b)| ((*a as f64 - b as f64).powf(2.0))).sum();
     //let mut mse: f64 = (mse as f64).powf(2.0);
     //println!{"{}, {}", target.len(), image.len()}
-
     return (mse / image.len() as f64);
 }
